backend/requirements.txt
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9